✅ components 통합 및 경로 수정
